Add warning for avoiding misusing
Added a system warning when taskType is specified without the corresponding Roborazzi plugin configuration. This is to prevent potential complications in screenshot testing as it won't allow Roborazzi plugin to generate the test report. This warning advises users to ensure proper setup in gradle.properties or via Gradle tasks for optimal functionality. Also refactored code to enhance readability by replacing multiple references to roborazziOptions.taskType with a taskType variable.