Add captureScreenRoboImage for capturing screen images, including dialogs
A new functionality has been added in the `Roborazzi.kt` file to capture screen images that include dialogs. The methods `captureScreenRoboImage()` have been introduced. This change allows for capturing full-screen images by traversing the active roots of the application and collecting both screen and dialog components into a Bitmap for output. This was not previously possible, as the captured images would not include dialogs, missing important interface elements in the testing context.

Furthermore, a new test case `DialogTest.kt` has been added demonstrating the usage of these new capabilities. This is crucial in evaluating the effectiveness of the screen capture functionality we have introduced."